+++ config.py
@@ -17,4 +17,6 @@ CATEGORY_NUMBER_DOWNLOAD_ITEMS = os.getenv('CATEGORY_NUMBER_DOWNLOAD_ITEMS', 100
 MAX_ITEM_COUNT_OFFER = os.getenv('MAX_ITEM_COUNT_OFFER', 10)
 MAX_ITEM_PAGE_OFFER = os.getenv('MAX_ITEM_PAGE_OFFER', 10)
 THROTTLING_SECONDS = os.getenv('THROTTLING_SECONDS', 2)
+CELERY_BROKER_URL =  os.getenv('CELERY_BROKER_URL', 'redis://localhost:6379/1')
+CELERY_RESULT_BACKEND = os.getenv('CELERY_RESULT_BACKEND', 'redis://localhost:6379/1')
 DEBUG = os.getenv('DEBUG', True)
