--- init_services/celery_init.py
@@ -1,21 +1,0 @@
-from celery import Celery
-
-CELERY_BROKER_URL = 'redis://localhost:6379/1'
-CELERY_RESULT_BACKEND = 'redis://localhost:6379/1'
-
-
-def make_celery(app):
-    celery = Celery(
-        app.import_name,
-        backend=CELERY_RESULT_BACKEND,
-        broker=CELERY_BROKER_URL
-    )
-    celery.conf.update(app.config)
-
-    class ContextTask(celery.Task):
-        def __call__(self, *args, **kwargs):
-            with app.app_context():
-                return self.run(*args, **kwargs)
-
-    celery.Task = ContextTask
-    return celery
