+++ main.py
@@ -1,20 +1,12 @@
 import config
 import init_server as init
-from init_services.celery_init import make_celery
-from init_services.flask_init import make_flask
-
-
-app = make_flask()
-celery = make_celery(app)
-
-
-@app.route('/')
-def hello_world():
-    return 'Hello, PA API Amazon Server!\n'
+import init_services.celery_services as celery_services
+from init_services.factory_services import *
 
 
 if __name__ == '__main__':
     if init.init_server():
+        app = init_services(celery=celery_services.celery_app)
         app.run(host='0.0.0.0', debug=config.DEBUG)
     else:
         print("###### THERE WAS AN ERROR, CHECK THE LOGS ######")
